# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a URL shortener service.",
        "title": "URL shortening API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/data/shorten": {
            "post": {
                "description": "generate shortened urls",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "generate shortened urls",
                "parameters": [
                    {
                        "description": "Add URL for shortening",
                        "name": "longUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LongUrl"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ShortenedUrl"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LongUrl": {
            "type": "object",
            "properties": {
                "longUrl": {
                    "type": "string"
                }
            }
        },
        "models.ShortenedUrl": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "longUrl": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                }
            }
        },
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}